separated System_R and System_k, removed ase and pythtb from github testing
