+++ tests/common_systems.py
@@ -314,12 +314,17 @@ def system_GaAs_tb_wcc_ws():
 
 
 # Haldane model from TBmodels
-model_tbmodels_Haldane = wb_models.Haldane_tbm(delta=0.2, hop1=-1.0, hop2=0.15)
 
 
 @pytest.fixture(scope="session")
 def system_Haldane_TBmodels():
     # Load system
+    try:
+        import tbmodels
+        tbmodels  # just to avoid F401
+    except (ImportError, ModuleNotFoundError):
+        pytest.xfail("failed to import tbmodels")
+    model_tbmodels_Haldane = wb_models.Haldane_tbm(delta=0.2, hop1=-1.0, hop2=0.15)
     system = wberri.system.System_TBmodels(model_tbmodels_Haldane, berry=True)
     system.set_symmetry(["C3z"])
     return system
@@ -328,6 +333,12 @@ def system_Haldane_TBmodels():
 @pytest.fixture(scope="session")
 def system_Haldane_TBmodels_internal():
     # Load system
+    try:
+        import tbmodels
+        tbmodels  # just to avoid F401
+    except (ImportError, ModuleNotFoundError):
+        pytest.xfail("failed to import tbmodels")
+    model_tbmodels_Haldane = wb_models.Haldane_tbm(delta=0.2, hop1=-1.0, hop2=0.15)
     system = wberri.system.System_TBmodels(model_tbmodels_Haldane, berry=False)
     system.set_symmetry(["C3z"])
     return system
