+++ wannierberri/system/system_w90.py
@@ -16,11 +16,11 @@ import numpy as np
 import functools
 import multiprocessing
 from ..__utility import iterate3dpm, real_recip_lattice, fourier_q_to_R
-from .system import System
+from .system_R import System_R
 from .w90_files import Wannier90data
 
 
-class System_w90(System):
+class System_w90(System_R):
     """
     System initialized from the Wannier functions generated by `Wannier90 <http://wannier.org>`__ code.
     Reads the ``.chk``, ``.eig`` and optionally ``.mmn``, ``.spn``, ``.uHu``, ``.sIu``, and ``.sHu`` files
